# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BSV blockchain indexer API for transaction outputs, blocks, and related data",
        "title": "1Sat Indexer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/shruggr/1sat-indexer"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "5.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v5/own/{owner}/balance": {
            "get": {
                "description": "Get the satoshi balance for a specific owner",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "owners"
                ],
                "summary": "Get owner balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Owner identifier (address, pubkey, or script hash)",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Refresh spend data from blockchain",
                        "name": "refresh",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Balance information",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v5/own/{owner}/txos": {
            "get": {
                "description": "Get transaction outputs owned by a specific owner (address/pubkey/script hash)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "owners"
                ],
                "summary": "Get owner TXOs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Owner identifier (address, pubkey, or script hash)",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Refresh owner data from blockchain",
                        "name": "refresh",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of tags to include (use * for all indexed tags)",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Starting score for pagination",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Reverse order",
                        "name": "rev",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include TXO data",
                        "name": "txo",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include script data",
                        "name": "script",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include spend information",
                        "name": "spend",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "Filter for unspent outputs only",
                        "name": "unspent",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/idx.Txo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v5/own/{owner}/utxos": {
            "get": {
                "description": "Get transaction outputs owned by a specific owner (address/pubkey/script hash)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "owners"
                ],
                "summary": "Get owner TXOs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Owner identifier (address, pubkey, or script hash)",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Refresh owner data from blockchain",
                        "name": "refresh",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of tags to include (use * for all indexed tags)",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Starting score for pagination",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Reverse order",
                        "name": "rev",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include TXO data",
                        "name": "txo",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include script data",
                        "name": "script",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include spend information",
                        "name": "spend",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "Filter for unspent outputs only",
                        "name": "unspent",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/idx.Txo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v5/spends": {
            "post": {
                "description": "Get spend information for multiple transaction outputs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spends"
                ],
                "summary": "Get multiple spend records",
                "parameters": [
                    {
                        "description": "Array of outpoints",
                        "name": "outpoints",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Spend information array",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v5/spends/{outpoint}": {
            "get": {
                "description": "Get spend information for a transaction output by outpoint",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spends"
                ],
                "summary": "Get spend information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction outpoint (txid_vout)",
                        "name": "outpoint",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Spend information",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v5/tag/{tag}": {
            "get": {
                "description": "Search for transaction outputs by tag",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Get TXOs by tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag name",
                        "name": "tag",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of tags to include (use * for all indexed tags)",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Starting score for pagination",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Reverse order",
                        "name": "rev",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include TXO data",
                        "name": "txo",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include script data",
                        "name": "script",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/idx.Txo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "evt.Event": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "idx.IndexData": {
            "type": "object",
            "properties": {
                "data": {},
                "deps": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/evt.Event"
                    }
                }
            }
        },
        "idx.Txo": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/idx.IndexData"
                    }
                },
                "events": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "height": {
                    "type": "integer"
                },
                "idx": {
                    "type": "integer"
                },
                "outpoint": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "owners": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "satoshis": {
                    "type": "integer"
                },
                "score": {
                    "type": "number"
                },
                "script": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "spend": {
                    "type": "string"
                }
            }
        }
    }
}